+++ src/backend/distributed/commands/index.c
@@ -173,6 +173,7 @@ PreprocessIndexStmt(Node *node, const char *createIndexCommand,
 														   namespaceName);
 	}
 
+	Oid relationOid = relation->rd_id;
 	table_close(relation, NoLock);
 
 	Oid relationId = CreateIndexStmtGetRelationId(createIndexStatement);
@@ -197,7 +198,7 @@ PreprocessIndexStmt(Node *node, const char *createIndexCommand,
 		 * it on a copy not to interfere with standard process utility.
 		 */
 		IndexStmt *copyCreateIndexStatement =
-			transformIndexStmt(relation->rd_id, copyObject(createIndexStatement),
+			transformIndexStmt(relationOid, copyObject(createIndexStatement),
 							   createIndexCommand);
 
 		/* ensure we copy string into proper context */