+++ src/backend/distributed/sql/citus--14.0-1--15.0-1.sql
@@ -1,0 +1,2 @@
+-- citus--14.0-1--15.0-1
+-- bump version to 15.0-1