+++ src/test/regress/sql/multi_extension.sql
@@ -766,6 +766,16 @@ SELECT * FROM multi_extension.print_extension_changes();
 ALTER EXTENSION citus UPDATE TO '14.0-1';
 SELECT * FROM multi_extension.print_extension_changes();
 
+-- Test downgrade to 14.0-1 from 15.0-1
+ALTER EXTENSION citus UPDATE TO '15.0-1';
+ALTER EXTENSION citus UPDATE TO '14.0-1';
+-- Should be empty result since upgrade+downgrade should be a no-op
+SELECT * FROM multi_extension.print_extension_changes();
+
+-- Snapshot of state at 15.0-1
+ALTER EXTENSION citus UPDATE TO '15.0-1';
+SELECT * FROM multi_extension.print_extension_changes();
+
 DROP TABLE multi_extension.prev_objects, multi_extension.extension_diff;
 
 -- show running version